requirements.txt
yfinance
phidata==2.7.10
groq
python-dotenv
duckduckgo-search
fastapi
uvicorn
utils
openai
youtube_transcript_api
